+++ lib/src/factory_impl.dart
@@ -1,12 +1,26 @@
 import 'dart:async';
 import 'dart:convert';
 import 'dart:html' as html;
+import 'package:js/js.dart';
 import 'package:webrtc_interface/webrtc_interface.dart';
 
 import 'media_recorder_impl.dart';
 import 'media_stream_impl.dart';
 import 'navigator_impl.dart';
 import 'rtc_peerconnection_impl.dart';
+import 'rtc_rtp_capailities_imp.dart';
+
+@JS('RTCRtpSender')
+@anonymous
+class RTCRtpSenderJs {
+  external static Object getCapabilities(String kind);
+}
+
+@JS('RTCRtpReceiver')
+@anonymous
+class RTCRtpReceiverJs {
+  external static Object getCapabilities(String kind);
+}
 
 class RTCFactoryWeb extends RTCFactory {
   RTCFactoryWeb._internal();
@@ -47,6 +61,18 @@ class RTCFactoryWeb extends RTCFactory {
 
   @override
   Navigator get navigator => NavigatorWeb();
+
+  @override
+  Future<RTCRtpCapabilities> getRtpReceiverCapabilities(String kind) async {
+    var caps = RTCRtpReceiverJs.getCapabilities(kind);
+    return RTCRtpCapabilitiesWeb.fromJsObject(caps);
+  }
+
+  @override
+  Future<RTCRtpCapabilities> getRtpSenderCapabilities(String kind) async {
+    var caps = RTCRtpSenderJs.getCapabilities(kind);
+    return RTCRtpCapabilitiesWeb.fromJsObject(caps);
+  }
 }
 
 Future<RTCPeerConnection> createPeerConnection(
@@ -58,6 +84,14 @@ Future<RTCPeerConnection> createPeerConnection(
 
 Future<MediaStream> createLocalMediaStream(String label) {
   return RTCFactoryWeb.instance.createLocalMediaStream(label);
+}
+
+Future<RTCRtpCapabilities> getRtpReceiverCapabilities(String kind) async {
+  return RTCFactoryWeb.instance.getRtpReceiverCapabilities(kind);
+}
+
+Future<RTCRtpCapabilities> getRtpSenderCapabilities(String kind) async {
+  return RTCFactoryWeb.instance.getRtpSenderCapabilities(kind);
 }
 
 MediaRecorder mediaRecorder() {